+++ config/webpack.config.js
@@ -12,7 +12,6 @@ require('babel-polyfill');
 const timestamp = new Date().getTime();
 
 const globalEntryFiles = {
-  'no-react': './src/js/no-react-entry.js',
   style: './src/sass/style.scss',
   vendor: [
     './src/platform/polyfills',
