--- src/js/facility-locator/store/index.js
@@ -1,4 +1,0 @@
-import createCommonStore from '../../common/store';
-import reducer from '../reducers';
-
-export const store = createCommonStore(reducer);
