+++ src/platform/site-wide/feedback/actions/index.js
@@ -1,5 +1,5 @@
-import { apiRequest } from '../../common/helpers/api';
-import { isValidEmail } from '../../common/utils/validations';
+import { apiRequest } from '../../../../js/common/helpers/api';
+import { isValidEmail } from '../../../../js/common/utils/validations';
 
 export const REVEAL_FORM = 'REVEAL_FORM';
 export const SET_FORM_VALUES = 'SET_FORM_VALUES';
