+++ src/platform/site-wide/index.js
@@ -3,15 +3,15 @@
  * @module platform/site-wide
  */
 
-// This will move later, mostly likely
+// This will move later, most likely
 import '../../js/common/sentry.js';
 
-import '../../js/legacy/menu';  // Used in the footer.
-import '../../js/common/usa-banner-toggle';
-import '../../js/common/utils/accessible-VCL-modal';
-import addMenuListeners from '../../js/common/utils/accessible-menus';
+import './legacy/menu';  // Used in the footer.
+import './usa-banner-toggle';
+import './accessible-VCL-modal';
+import addMenuListeners from './accessible-menus';
 import createLoginWidget from '../../js/login/login-entry';
-import createFeedbackWidget from '../../js/feedback/feedback-entry';
+import createFeedbackWidget from './feedback/feedback-entry';
 
 /**
  * Start up the site-wide components that live on every page, like 
@@ -24,7 +24,7 @@ export default function startSitewideComponents(commonStore) {
 
   // New navigation menu
   if (document.querySelector('#vetnav')) {
-    require('../../js/legacy/mega-menu.js');
+    require('./legacy/mega-menu.js');
   }
 
   // Prevent some browsers from changing the value when scrolling while hovering
