+++ src/platform/site-wide/usa-banner-toggle.js
@@ -1,3 +1,7 @@
+/**
+ * Adds event handler to show/hide the USA site banner
+ */
+
 function initBanner() {
 
   const toggleButton = document.querySelector('#usa-banner-toggle');
