+++ src/sass/shared-variables.scss
@@ -1,5 +1,5 @@
 // This file should import the variables that are common between style.scss
-// and the individual react apps (plus no-react)
+// and the individual react apps (plus static-pages)
 // You should not add actual variables in here, but if you create a new file for variables
 // that you want to use in all apps, it would be imported in here
 @charset "UTF-8";
