+++ src/sass/static-pages.scss
