+++ test/gibct/containers/SearchPage.unit.spec.js
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import SkinDeep from 'skin-deep';
 import sinon from 'sinon';
 
-import createCommonStore from '../../../src/js/common/store';
+import createCommonStore from '../../../src/platform/startup/store';
 import { SearchPage } from '../../../src/js/gi/containers/SearchPage';
 import reducer from '../../../src/js/gi/reducers';
