+++ test/user-profile/containers/UserProfileApp.unit.spec.jsx
@@ -3,7 +3,7 @@ import SkinDeep from 'skin-deep';
 import { expect } from 'chai';
 
 import UserProfileApp from '../../../src/js/user-profile/containers/UserProfileApp';
-import createCommonStore from '../../../src/js/common/store';
+import createCommonStore from '../../../src/platform/startup/store';
 
 const store = createCommonStore();
